refactor folders
remove misc module
add progress bar
