+++ mupl/uploader/handler.py
@@ -1,43 +1,31 @@
-import asyncio
 import logging
-import os
-import time
-from pathlib import Path
 from typing import List, Optional, Dict
 
-from uploader.file_validator import FileProcesser
-from uploader.http import RequestError
-from uploader.http.client import HTTPClient
-from uploader.image_validator import ImageProcessor
-from uploader.utils.config import (
-    NUMBER_THREADS,
+from mupl.file_validator import FileProcesser
+from mupl.http import RequestError
+from mupl.http.client import HTTPClient
+from mupl.image_validator import ImageProcessor
+from mupl.utils.config import (
+    VERBOSE,
     config,
     mangadex_api_url,
-    RATELIMIT_TIME,
     UPLOAD_RETRY,
 )
-from uploader.utils.misc import create_new_event_loop, flatten
 
 logger = logging.getLogger("md_uploader")
 
 
-class ChapterUploader:
+class ChapterUploaderHandler:
     def __init__(
         self,
         http_client: "HTTPClient",
         file_name_obj: "FileProcesser",
-        names_to_ids: "dict",
         failed_uploads: "list",
-        threaded: "bool",
     ):
         self.http_client = http_client
         self.file_name_obj = file_name_obj
         self.to_upload = self.file_name_obj.to_upload
-        self.names_to_ids = names_to_ids
         self.failed_uploads = failed_uploads
-        self.threaded = threaded
-        if NUMBER_THREADS <= 1:
-            self.threaded = False
         self.zip_name = self.to_upload.name
         self.zip_extension = self.to_upload.suffix
         self.folder_upload = False
@@ -46,9 +34,7 @@ class ChapterUploader:
             self.folder_upload = True
             self.zip_extension = None
 
-        self.uploaded_files_path = Path(config["paths"]["uploaded_files"])
         self.number_upload_retry = UPLOAD_RETRY
-        self.ratelimit_time = RATELIMIT_TIME
         self.md_upload_api_url = f"{mangadex_api_url}/upload"
 
         # Images to include with chapter commit
@@ -59,7 +45,6 @@ class ChapterUploader:
         self.image_uploader_process = ImageProcessor(
             self.file_name_obj, self.folder_upload
         )
-        self.myzip = self.image_uploader_process.myzip
 
     def _images_upload(self, image_batch: "Dict[str, bytes]"):
         """Upload the images"""
@@ -94,16 +79,16 @@ class ChapterUploader:
 
     def _upload_images(self, image_batch: "Dict[str, bytes]") -> "bool":
         """Try to upload every 10 (default) images to the upload session."""
-        # No images to upload
         if not image_batch:
             return True
 
         successful_upload_message = "Success: Uploaded page {}, size: {} MB."
 
         image_batch_list = list(image_batch.keys())
-        print(
-            f"Uploading images {int(image_batch_list[0]) + 1} to {int(image_batch_list[-1]) + 1}."
-        )
+        if VERBOSE:
+            print(
+                f"Uploading images {int(image_batch_list[0]) + 1} to {int(image_batch_list[-1]) + 1}."
+            )
         logger.debug(
             f"Uploading images {int(image_batch_list[0]) + 1} to {int(image_batch_list[-1]) + 1}."
         )
@@ -141,17 +126,20 @@ class ChapterUploader:
                 if converted_format is not None:
                     formatted_name_message += f" (converted to {converted_format})"
 
-                print(
-                    successful_upload_message.format(
-                        formatted_name_message, round(file_size * 0.00000095367432, 2)
+                if VERBOSE:
+                    print(
+                        successful_upload_message.format(
+                            formatted_name_message,
+                            round(file_size * 0.00000095367432, 2),
+                        )
                     )
-                )
 
             # Length of images array returned from the api is the same as the array sent to the api
             if len(successful_upload_data) == len(image_batch):
                 logger.info(
                     f"Uploaded images {int(image_batch_list[0]) + 1} to {int(image_batch_list[-1]) + 1}."
                 )
+                self.tqdm.update(len(image_batch_list))
                 return False
             else:
                 # Update the images to upload dictionary with the images that failed
@@ -282,155 +270,3 @@ class ChapterUploader:
         self.remove_upload_session()
         self.failed_uploads.append(self.to_upload)
         return False
-
-    def _move_files(self):
-        """Move the uploaded chapters to a different folder."""
-        self.uploaded_files_path.mkdir(parents=True, exist_ok=True)
-        # Folders don't have an extension
-        if self.folder_upload:
-            zip_name = self.zip_name
-        else:
-            zip_name = self.zip_name.rsplit(".", 1)[0]
-        zip_extension = self.zip_extension or ""
-        zip_path_str = f"{zip_name}{zip_extension}"
-        version = 1
-
-        # If a file/folder with that name exists already in the uploaded files path
-        # Add a version number to the end before moving
-        while True:
-            version += 1
-            if zip_path_str in os.listdir(self.uploaded_files_path):
-                if self.folder_upload:
-                    zip_name_unformat = self.zip_name
-                else:
-                    zip_name_unformat = self.zip_name.rsplit(".", 1)[0]
-
-                zip_name = f"{zip_name_unformat}{{v{version}}}"
-                zip_path_str = f"{zip_name}{zip_extension}"
-                continue
-            else:
-                break
-
-        new_uploaded_zip_path = self.to_upload.rename(
-            os.path.join(self.uploaded_files_path, f"{zip_name}{zip_extension}")
-        )
-        logger.debug(f"Moved {self.to_upload} to {new_uploaded_zip_path}.")
-
-    async def process_images_upload(self, images_array):
-        """Start uploading the images, threaded."""
-        images_to_upload = self.image_uploader_process.get_images_to_upload(
-            images_array
-        )
-        failed = self._upload_images(images_to_upload)
-        if failed:
-            self.failed_image_upload = True
-            asyncio.get_running_loop().close()
-
-    def run_threaded_uploader(self, spliced_images):
-        """Run the threads for upload."""
-        tasks = []
-
-        loop = create_new_event_loop()
-        for images_to_upload in spliced_images:
-            task = self.process_images_upload(images_to_upload)
-            tasks.append(task)
-
-        gathered = asyncio.gather(*tasks)
-
-        try:
-            loop.run_until_complete(gathered)
-        except KeyboardInterrupt as e:
-            print("Caught keyboard interrupt. Canceling tasks...")
-            gathered.cancel()
-            self.failed_image_upload = True
-
-    def run_image_uploader(self, images):
-        """Run the image uploader ."""
-        for images_array in images:
-            images_to_upload = self.image_uploader_process.get_images_to_upload(
-                images_array
-            )
-            failed = self._upload_images(images_to_upload)
-            if failed:
-                self.failed_image_upload = True
-
-            # Don't upload rest of the chapter's images if the images before failed
-            if self.failed_image_upload:
-                break
-
-    def start_chapter_upload(self):
-        """Process the zip for uploading."""
-        upload_details = (
-            f"Manga id: {self.file_name_obj.manga_series}, "
-            f"chapter: {self.file_name_obj.chapter_number}, "
-            f"volume: {self.file_name_obj.volume_number}, "
-            f"title: {self.file_name_obj.chapter_title}, "
-            f"language: {self.file_name_obj.language}, "
-            f"groups: {self.file_name_obj.groups}, "
-            f"publish on: {self.file_name_obj.publish_date}."
-        )
-        logger.info(f"Uploading chapter: {upload_details}")
-        print(upload_details)
-
-        if not self.image_uploader_process.valid_images_to_upload:
-            no_valid_images_found_error_message = (
-                f"{self.zip_name} has no valid images to upload, skipping."
-            )
-            print(no_valid_images_found_error_message)
-            logger.error(no_valid_images_found_error_message)
-            self.failed_uploads.append(self.to_upload)
-            return
-
-        self.http_client.login()
-
-        upload_session_response_json = self._create_upload_session()
-        if upload_session_response_json is None:
-            time.sleep(self.ratelimit_time)
-            return
-
-        self.upload_session_id = upload_session_response_json["data"]["id"]
-
-        upload_session_id_message = (
-            f"Created upload session: {self.upload_session_id}, {self.zip_name}."
-        )
-        logger.info(upload_session_id_message)
-        print(upload_session_id_message)
-        print(
-            f"{len(flatten(self.image_uploader_process.valid_images_to_upload))} images to upload."
-        )
-
-        spliced_images_list = [
-            self.image_uploader_process.valid_images_to_upload[
-                elem : elem + NUMBER_THREADS
-            ]
-            for elem in range(
-                0,
-                len(self.image_uploader_process.valid_images_to_upload),
-                NUMBER_THREADS,
-            )
-        ]
-
-        if self.threaded:
-            print("Running threaded uploader.")
-            for spliced_images in spliced_images_list:
-                self.run_threaded_uploader(spliced_images)
-                if self.failed_image_upload:
-                    break
-        else:
-            print("Running non-threaded uploader.")
-            self.run_image_uploader(self.image_uploader_process.valid_images_to_upload)
-
-        if not self.folder_upload:
-            self.myzip.close()
-
-        # Skip chapter upload and delete upload session
-        if self.failed_image_upload:
-            failed_image_upload_message = f"Deleting draft due to failed image upload: {self.upload_session_id}, {self.zip_name}."
-            print(failed_image_upload_message)
-            logger.error(failed_image_upload_message)
-            self.remove_upload_session()
-            self.failed_uploads.append(self.to_upload)
-            return
-
-        logger.info("Uploaded all of the chapter's images.")
-        self._commit_chapter()
