+++ mupl/utils/__init__.py
@@ -1,0 +1,1 @@
+from mupl.utils.logs import setup_logs
