--- uploader/utils/__init__.py
@@ -1,1 +1,0 @@
-from uploader.utils.logs import setup_logs
